Spring now reads user credentials from database

+++ artsholland-api/src/main/java/org/waag/ah/spring/service/CRUDService.java
@@ -1,6 +1,5 @@
 package org.waag.ah.spring.service;
 
-import java.lang.reflect.Field;
 import java.lang.reflect.ParameterizedType;
 import java.util.Collection;
 
@@ -9,6 +8,7 @@ import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.Query;
 
+import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.waag.ah.model.DbObject;
@@ -17,7 +17,7 @@ import org.waag.ah.model.DbObject;
 @Transactional
 public abstract class CRUDService<T extends DbObject> {
 
-	// TODO: user Spring transactions
+	// TODO: use Spring transactions
 	
 	protected Class<T> type;
 	
@@ -35,19 +35,27 @@ public abstract class CRUDService<T extends DbObject> {
 		entityManager.merge(object); 
 	}
 
-	public T read(long id) {		
-    T dbObject = entityManager.find(type, id);
+	public T read(Object primaryKey) {		
+    T dbObject = entityManager.find(type, primaryKey);
     return dbObject;
 	}
 
-	// TODO: update all fields with reflection?
-	public abstract void update(T object);
-
-	public void delete(T object) {		
-    T dbObject = entityManager.find(type, object.getId());
-    entityManager.remove(dbObject); 
+	public void update(T object) {
+		T dbObject = entityManager.find(type, object.getId());		
+		if (dbObject != null) {
+			BeanUtils.copyProperties(object, dbObject);
+		}		
+		entityManager.merge(dbObject);
 	}
 
+	public void delete(Object primaryKey) {		
+    T dbObject = entityManager.find(type, primaryKey);
+    if (dbObject != null) {
+    	entityManager.remove(dbObject);
+    }
+	}
+
+	@SuppressWarnings("unchecked")
 	public Collection<T> findAll() {		
 		Query query = entityManager.createQuery("SELECT e FROM " + type.getSimpleName() + " e");
 		return (Collection<T>) query.getResultList();
